alias-cbu-relayer/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
pydantic==2.9.2
starknet-py==0.21.0
python-dotenv==1.0.1
